move dataset generation to separate file, remove useless  function
