+++ index.html
@@ -36,10 +36,10 @@
                 <input
                   type="radio"
                   name="target"
-                  id="Vanguard_-_Archer"
+                  id="Vanguard / Archer"
                   class="form-check-input flex-shrink-0"
                 />
-                <label for="Vanguard_-_Archer">Vanguard / Archer</label>
+                <label for="Vanguard / Archer">Vanguard / Archer</label>
               </div>
               <small>+0%</small>
               <div class="labelled-input">
